python/croatianbot/requirements.txt
google-cloud-secret-manager
google-cloud-storage